+++ reapy/core/reaper/__init__.py
@@ -1,3 +1,3 @@
-from . import audio
+from . import audio, midi
 from .reaper import *
 from .defer import defer, at_exit