+++ reapy/core/reaper/midi.py
@@ -1,0 +1,6 @@
+import reapy.reascript_api as RPR
+
+
+def reinit():
+    """Reset all MIDI devices."""
+    RPR.midi_reinit()